Auth improved

+++ AppointmentsService/Controllers/AppointmentsController.cs
@@ -7,12 +7,14 @@ using CommonData.Constants;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.Tokens;
 using Serilog;
 
 namespace AppointmentsService.Controllers
 {
     [ApiController]
     [Route("[controller]")]
+    [Authorize]
     public class AppointmentsController : ControllerBase
     {
         private readonly IMapper _mapper;
@@ -24,7 +26,6 @@ namespace AppointmentsService.Controllers
         }
 
         [HttpGet]
-        //[Authorize]
         public async Task<ActionResult<IEnumerable<ClientAppointment>>> GetAppointments([FromQuery] int pageNumber, 
                                                                                         [FromQuery] int pageSize,
                                                                                         [FromQuery] DateTime? date, 
@@ -34,19 +35,19 @@ namespace AppointmentsService.Controllers
                                                                                         [FromQuery] Guid? officeId, 
                                                                                         [FromQuery] Guid? patientId)
         {
-            //if (User.IsInRole(Roles.Patient))
-            //{
-            //    var userId = User.Claims.FirstOrDefault(c => c.Type == "Id");
-                
-            //    if(userId == null)
-            //    {
-            //        return Forbid();
-            //    }
-            //    else
-            //    {
-            //        patientId = Guid.Parse(userId.Value);
-            //    }
-            //}
+            if (User.IsInRole(Roles.Patient))
+            {
+                var userId = User.Claims.FirstOrDefault(c => c.Type == "Id");
+
+                if (userId == null || userId.Value.IsNullOrEmpty())
+                {
+                    return Forbid();
+                }
+                else
+                {
+                    patientId = Guid.Parse(userId.Value);
+                }
+            }
 
             var query = new GetAppointmentsQuery()
             {
@@ -64,7 +65,7 @@ namespace AppointmentsService.Controllers
 
             var clientAppointments = _mapper.Map<IEnumerable<ClientAppointment>>(appointments);
 
-            return new(clientAppointments);
+            return Ok(clientAppointments);
         }
 
         [HttpGet("{id:Guid}")]
@@ -77,10 +78,10 @@ namespace AppointmentsService.Controllers
 
             DbAppointment appointments = await _mediator.Send(query);
 
-            //if (User.IsInRole(Roles.Patient) && !UserIdMatches(appointments.PatientId))
-            //{
-            //    return Forbid();
-            //}
+            if (User.IsInRole(Roles.Patient) && !UserIdMatches(appointments.PatientId))
+            {
+                return Forbid();
+            }
 
             var clientAppointment = _mapper.Map<ClientAppointment>(appointments);
 
@@ -88,7 +89,6 @@ namespace AppointmentsService.Controllers
         }
 
         [HttpPost]
-        //[Authorize("appointments.edit")]
         public async Task<ActionResult> AddAppointment([FromBody] ClientAppointment appointment)
         {
             var dbAppointment = _mapper.Map<DbAppointment>(appointment);
@@ -96,19 +96,18 @@ namespace AppointmentsService.Controllers
             // todo: how to return guid or object?
             await _mediator.Send(new AddAppointmentCommand() { Appointment = dbAppointment });
 
-            Log.Information("Appointment created => {@record}", (dbAppointment.Id, dbAppointment.ServiceName, dbAppointment.OfficeAddress));
+            Log.Information("Appointment created: {@record}", (dbAppointment.Id, dbAppointment.ServiceName, dbAppointment.OfficeAddress));
 
             return Ok();
         }
 
         [HttpPut]
-        //[Authorize("appointments.edit")]
         public async Task<ActionResult<ClientAppointment>> UpdateAppointment([FromBody] ClientAppointment clientAppointment)
         {
-            //if (!UserIdMatches(clientAppointment.PatientId) && !User.IsInRole(Roles.Receptionist) && !User.IsInRole(Roles.Admin))
-            //{
-            //    return Forbid();
-            //}
+            if (!UserIdMatches(clientAppointment.PatientId) && !User.IsInRole(Roles.Receptionist))
+            {
+                return Forbid();
+            }
 
             var dbAppointment = _mapper.Map<DbAppointment>(clientAppointment);
 
@@ -125,13 +124,12 @@ namespace AppointmentsService.Controllers
         }
 
         [HttpPut("approve/{id:Guid}")]
-        //[Authorize("appointments.edit")]
         public async Task<ActionResult<ClientAppointment>> ChangeAppointmentApproval(Guid id, [FromQuery] bool approved)
         {
-            //if (!User.IsInRole(Roles.Receptionist))
-            //{
-            //    return Forbid();
-            //}
+            if (!User.IsInRole(Roles.Receptionist))
+            {
+                return Forbid();
+            }
 
             var command = new ChangeAppointmentApprovalCommand()
             {
@@ -145,9 +143,13 @@ namespace AppointmentsService.Controllers
         }
 
         [HttpDelete("{id:Guid}")]
-        //[Authorize("appointments.edit")]
         public async Task<ActionResult> DeleteAppointment(Guid id)
         {
+            if(!User.IsInRole(Roles.Receptionist))
+            {
+                return Forbid();
+            }
+
             var command = new DeleteAppointmentCommand()
             {
                 Id = id
@@ -155,7 +157,7 @@ namespace AppointmentsService.Controllers
 
             await _mediator.Send(command);
 
-            Log.Information("Appointment deleted => {@record}", id);
+            Log.Information("Appointment deleted: {@record}", id);
 
             return NoContent();
         }
