+++ OfficesService/Controllers/OfficesController.cs
@@ -1,4 +1,5 @@
 using AutoMapper;
+using CommonData.Constants;
 using CommonData.Messages;
 using MassTransit;
 using Microsoft.AspNetCore.Authorization;
@@ -12,7 +13,6 @@ namespace OfficesService.Controllers
 {
     [ApiController]
     [Route("[controller]")]
-    [Authorize("offices.edit")]
     public class OfficesController : ControllerBase
     {
         private readonly IRepository<DbOfficeModel> _officesRepository;
@@ -24,6 +24,7 @@ namespace OfficesService.Controllers
         }
 
         [HttpGet]
+        [AllowAnonymous]
         public async Task<ActionResult<IEnumerable<ClientOfficeModel>>> GetOffices()
         {
             var dbOffices = await _officesRepository.GetAll();
@@ -34,6 +35,7 @@ namespace OfficesService.Controllers
         }
 
         [HttpGet("{id:Guid}")]
+        [AllowAnonymous]
         public async Task<ActionResult<ClientOfficeModel>> GetOffice(Guid id)
         {
             var dbOffice = await _officesRepository.Get(id);
@@ -44,23 +46,35 @@ namespace OfficesService.Controllers
         }
 
         [HttpDelete("{id:Guid}")]
+        [Authorize]
         public async Task<ActionResult> DeleteOffice(Guid id)
         {
+            if(!User.IsInRole(Roles.Receptionist))
+            {
+                return Forbid();
+            }
+
             await _officesRepository.Delete(id);
 
-            Log.Information("Office deleted => {@dbOffice}", id);
+            Log.Information("Office deleted: Id = {@dbOffice}", id);
 
             return NoContent();
         }
 
         [HttpPost]
+        [Authorize]
         public async Task<ActionResult<ClientOfficeModel>> CreateOffice([FromBody] ClientOfficeModel office)
         {
+            if (!User.IsInRole(Roles.Receptionist))
+            {
+                return Forbid();
+            }
+
             var dbOffice = _mapper.Map<DbOfficeModel>(office);
 
             var addedOffice = await _officesRepository.Add(dbOffice);
 
-            Log.Information("Office created => {@addedOffice}", addedOffice);
+            Log.Information("Office created: {@addedOffice}", (addedOffice.Id, addedOffice.Address));
 
             var clientOffice = _mapper.Map<ClientOfficeModel>(addedOffice);
 
@@ -68,13 +82,19 @@ namespace OfficesService.Controllers
         }
 
         [HttpPut]
+        [Authorize]
         public async Task<ActionResult<ClientOfficeModel>> UpdateOffice([FromBody] ClientOfficeModel office)
         {
+            if (!User.IsInRole(Roles.Receptionist))
+            {
+                return Forbid();
+            }
+
             var dbOffice = _mapper.Map<DbOfficeModel>(office);
 
             var updatedOffice = await _officesRepository.Update(dbOffice);
 
-            Log.Information("Office updated => {@addedOffice}", updatedOffice);
+            Log.Information("Office updated: {@updatedOffice}", (updatedOffice.Id, updatedOffice.Address));
 
             var clientOffice = _mapper.Map<ClientOfficeModel>(updatedOffice);
